backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
gunicorn>=21.2.0
pytrends>=4.9.2
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
